+++ sphinx_gallery/__init__.py
@@ -4,7 +4,7 @@ Sphinx Gallery
 
 """
 import os
-__version__ = '0.4.0'
+__version__ = '0.5.0.dev0'
 
 
 def glr_path_static():